Sql bindings for Example Sum aggregator

+++ src/main/java/io/imply/druid/example/aggregator/ExampleSumAggregatorFactory.java
@@ -165,7 +165,7 @@ public class ExampleSumAggregatorFactory extends AggregatorFactory
   @Override
   public ColumnType getIntermediateType()
   {
-    return ColumnType.FLOAT;
+    return ColumnType.DOUBLE;
   }
 
   @Override
